Update quickstart.rst
Hi Regis,
I was trying for a long time to get the tutor starting with errors I have no idea about. 
Then I thought it must be the docker engine, when i ran the daemon manually then i got a permission denied
I know maybe I should configure docker better but as a pretty new person in all of this it might be handy to explicit mention it
When I performed the su command prior to tutor local quickstart it worked smoothly to install which is what is happening now. 
Make sure that you are running the docker engine with 
'systemctl docker start' 
and it might be needed that you are logged in with root with you su command